backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow
python-dotenv==1.0.1
django-filter==23.5
django-storages==1.14.2
drf-spectacular==0.27.1
django-import-export==3.3.6
django-background-tasks==1.2.5
boto3==1.34.69 